refactor: moved default cache handler to external
